Modify the test and change the location of the test
